帰ります

+++ Dxlib1/Game/Ghost.cpp
@@ -64,7 +64,7 @@ void Ghost::Update()
 	kX_ = indexX_;
 	kY_ = indexY_;
 
-	moveInterval_ = Field::CHIP_SIZE / speed_;
+	moveInterval_ = static_cast<int>(Field::CHIP_SIZE / speed_);
 
 	// 移動のインターバル
 	if (moveTimer_ % moveInterval_ == 0)
